Add integration tests for the instruments.
There were already a number of tests that asserted on various instrument failures but none (afaict) that actually ran a benchmark and asserted on the measurements.

Also, resolve an issue that made it impossible to attach a debugger to caliper if it launched a child process.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=49061426