Attempt #2 to make ExperimentingCaliperRun#measure less insane.
The best thing that has ocurred is that instead of calling System.exit() when the worker crashes, we throw a TrialFailedException.

Rather than using nio channels just extract the stream reading and writing
code to another class and move the LineConsumer logic into the measure
method.  This creates a similar structure to [] but I believe it to
be superior for a few reasons.

1. It doesn't use nio Pipes to implement non blocking reads/writes (instead
it is just normal BlockingQueues and poison pills) so there are fewer system
calls and copies involved (which I believe will be more efficient, though it
remains to be proved).
2. It doesn't use nio Selectors/Channels which are a pain to use and configure (reading a line of text in a non blocking manner off of a channel takes way too much custom code).
3. It saves a similar number of threads (the process waiting thread, the consumer thread)
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=51952916