Replace Guice with Dagger2 for dependency injection
Some Guice functionality, specifically the dynamic creation of
classes, does not work on Android. The code that Dagger2
generates will however work on Android so is a suitable
replacement.

Some issues/reduction in capability:

1. Both benchmark classes and ResultProcessor instances were created through the Guice injector which meant that they could potentially use @Inject and have access to any bindings available from that injector, although the available bindings are not defined as part of the public API. Dagger 2 has no dynamic capabilities like Guice and so in order to work with these changes they would need a public no-argument constructor.
2. There are a number of groups of classes that are configurable, e.g. Instrument and ResultProcessor. Custom versions of ResultProcessor are affected by the previous issue. Built in instances of ResultProcessor and Instrument are both managed in a similar way.

The available built in instances are accessible through a `Map<Class<? extends T>, Provider<T>>` (where T is either ResultProcessor or Instrument). The list of configured classes is then scanned, and each configured class is used to find the appropriate `Provider<T>` from the map, it is an error if it does not exist. The `Provider<T>` is then used to create the instance of that object. This uses the `@Provides(type = Type.MAP)` mechanism coupled with custom `@MapKey` annotations, one for each type. That means that adding new types of Instrument or ResultProcessor (built in ones that do not have a public no-argument constructor) requires adding new @Provides methods.
3. The `@MainScope` custom scope annotation is really an alias for the `@Singleton`. The reason why it is needed is that `@Singleton` components cannot depend on other `@Singleton` components, but a custom scoped component can depend on any number of other components in the same scope.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=105591245