Allow loading of benchmark classes when running caliper as one jar
When caliper starts worker processes that measure allocations it
adds a special instrumentation agent jar to the command line
that measures the allocations. That jar is also added to the
boot class path using -Xbootclasspath/a. It finds the agentjar
by searching for a jar on the classpath that has a Premain-Class
attribute in its manifest with a specific value.

When caliper is run as a collection of jars, then the
java-allocation-instrumenter-3.0.jar is found on the
classpath and used as the agentjar and added to the boot class
path. However, when caliper is run as a single jar, e.g. as
caliper/target/caliper-1.0-SNAPSHOT-all.jar then caliper itself
is used as the agentjar and added to the boot class path. That
means that caliper classes are loaded by the boot class loader
and so need to use the thread context class loader to load the
benchmark classes which are loaded by the user/application
class loader.

This contains the following changes:

* Replace usages of Class.forName(String) with a call to a new
  Util.forBenchmarkClass(String) method that uses the thread
  context class loader to load the classes.
* Add a test that builds a single jar for caliper and attempts
  to run a process that simply loads a class using the
  Util.forBenchmarkClass(String) method.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=102980233