Warm up the method invocation path and improve determinism detection in allocation instrument
When run on certain JVMs the baseline recorded in
MicrobenchmarkAllocationWorker has allocations that are not present
in the real measurements which was treated as being
non-deterministic. Examination of the differences in allocations
indicated that the difference was caused by the JVM choosing a
different path for the first couple of invocations compared with
the last. Invoking the benchmark method a couple of extra times
seemed to fix the problem.

While doing that the method for detecting non-determinism was
proven to be flaky, changing the number of methods called caused
some of the tests to fail. So, the detection was improved a
little to ensure more consistent behaviour. Basically, it
attempts to call the method up to 5 times and checks to make
sure that the allocations measured are the same from one call to
the next. Two consecutive calls with the same allocations is
enough for the benchmark to be treated as deterministic.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=103093972