Remove CaliperOptionsComponent
This removes the CaliperOptionsComponent and adds the
OptionsModule to MainComponent.

The BenchmarkClassModule no longer takes the `Class<?>` object for
the benchmark as a constructor parameter, instead it gets it
from a provider, qualified by the Running.BenchmarkClass
annotation.

When the BenchmarkClassModule is used in the MainComponent the
benchmark class object is provided by the newly added MainModule
which bridges the gap between what the OptionsModule provides
(CaliperOptions) and what the BenchmarkClassModule needs.

In the WorkerComponent the WorkerModule provides what
BenchmarkClassModule needs.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=105947793