Avoid having to pass a dummy class name into the OptionsModule
The OptionsModule is used in a number of different places, some
of which require a benchmark class name to be provided and some
of which do not use it at all. At the moment those places that
do not require it have to add a dummy value for it which is
slightly confusing but not a major issue as the arguments are
fixed so it is guaranteed that it will never cause confusing for
the users of caliper.

However, as part of a future change it will be necessary to pass
an arbitrary set of user supplied arguments through to the
OptionsModule in a situation where the benchmark class name is
not required and where providing a dummy class could cause
confusion to the users if there was a problem with the arguments
they specified.

This change allows the ParsedOptions class to be told whether it
requires a benchmark class name or not and produce an
appropriate error message in either case if incorrect arguments
are provided.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=107775316