Refactor to support running on Android/Dalvik platform
Involves separating out all the JVM specific functionality and Android specific functionality into two separate parts, a JvmPlatform and DalvikPlatform each of which extend Platform class. Any usages of JVM specific functionality is then replaced with a call to an appropriate method on Platform.

The correct Platform is selected by using the DalvikPlatform (provided as an `Optional<DalvikPlatform>`) if it is available and then falling back to JvmPlatform as the default. This avoids having to come up with some criteria on when to use JvmPlatform that will match all the environments within which Caliper is already used and already works.

Most of the usages of Platform are hidden away within a VmConfig which delegates a lot of its work to Platform. The reason why Platform functionality wasn't merged into VmConfig is simply that a VmConfig refers to a specific virtual machine (e.g. a JVM at a specific location) while a Platform refers to a class of virtual machines (e.g. all Android virtual machines).
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=107782449