Move more things around.
- Move caliper.config, caliper.options and caliper.platform to the caliper-runner module. Classes in those packages are only used by the runner.
- Simplify binding of the current Platform (there was no need for separate modules for JvmPlatform and DalvikPlatform).
- Move the DalvikPlatform and JvmPlatform classes into the caliper.platform package. They don't really need their own packages.
- Remove BenchmarkClassChecker[Test]. It was entirely unused other than its test and one of the methods on MainComponent was only there to support it.

This is partly in preparation for separating out a new component that contains all the bindings that are needed *before* the benchmark class has been inspected so that we can use it to create workers, etc. MainComponent currently contains both those configuration-related bindings and the bindings for the benchmark class itself.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=164465052