Add an ID to each experiment for use in communication between workers and the runner.
Also create ExperimentSpec, a serializable version of an Experiment that can be sent to a worker, and adapt TrialRequest to primarily be a wrapper around an ExperimentSpec. As part of this, change all WorkerRequests to contain a UUID for identifying the worker, rather than just TrialRequest.

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=180936811