Change how requests are sent to workers.
Rather than serializing and base64 encoding the request object and sending it on the command line, just send the worker's ID and port on the command line (allowing socket communication with the runner to be established), then send the request via the socket.

This partly just because it makes more sense, partly to make the command line a bit more readable (I want to include the command line in the header for each worker's log file to make diagnosing issues a bit easier) and partly because it moves things a little closer to a design where a single worker process could handle multiple requests (not something I have concrete plans to make happen in the near future, but something that I think might actually make sense all the same).

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=184176983