Refactor caliper.runner package, splitting it into more subpackages to more logically group things and reduce the number of files lumped together in one place.
Changes are almost entirely moves, with some visibility changes to classes and methods. There are also a couple changes to the Instrument API to avoid circular package dependencies.

Since it's a bit hard to see the structure of this change in diff form, the new package structure (not including subpackages that were already there) is:

- runner
  - experiment
  - instrument
  - resultprocessor
  - target
  - worker
    - benchmarkmodel
    - dryrun
    - trial

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=187491961