Rename DeviceService to Device.
Despite being a service, it is primarily an abstraction of a device that happens to be a service. It will be the device part of the (Device, Vm) pair that makes a Target, and Device reads much more nicely in most contexts.

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=190221934