Don't call stopped() from VmProcess.kill().
The main reason for this change is to fix an issue where one or more exceptions were thrown when you ctrl-C'd Caliper; the shutdown hook thread for each active worker process would run, which would call kill() on the process; if kill() then called stopped(), the listener would try to remove the shutdown hook for the process, which would throw an exception from Runtime.removeShutdownHook because shutdown was already in process.

This is also more consistent with Process.destroy(), which doesn't wait for the process to exit; as such it wasn't even necessarily true that the process was stopped when stopped() was called, just that we'd asked it to stop.

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=194088790