Allow users to not specify a device on the command line for Android benchmarks as well as for JVM benchmarks.
Basically, the default device will still be "local" in general, but when no device is specified and --worker-classpath-android *is* specified, the default device will instead be "android" (i.e. the single adb-connected device or emulator, failing if more than one is present).

Primarily, this means that users of android_benchmarks will not *always* need to provide a device on the command line.

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=194091808