Fix Caliper tests.
A recent change to the allocation instrumenter broke one of our allocation instrument tests. The test claimed to be expecting one object per rep, but was using a LinkedList and adding an Object to it each rep, which is two objects. It seems like this was somehow counteracted by the fact that something in the allocation recorder itself was using a LinkedList, and switching that to an ArrayList made the numbers show up as what you'd actually expect: two objects per rep, which failed the test.

In any case, I've changed the benchmark in question to actually only allocate one object per rep.

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=195102094