Sort the detailed allocations data
I was attempting to use this to understand the allocation difference between 2 scenarios in a benchmark.  The plan was to gather all allocations between the two runs and then use diff to compare.  However, the instability of the ConcurrentHashSet makes this very difficult, so this cl sorts the data prior to printing.

RELNOTES=When setting `instrument.allocation.options.trackAllocations=true` the logged allocation data will be deterministically sorted.
PiperOrigin-RevId: 351892153